+++ cli/utils/build_log.go
@@ -5,14 +5,15 @@ package utils
  * I use this in testing to ensure that the expected outcomes are met.
  */
 type BuildLog struct {
-	imageHash       string
-	version         string
-	hashExists      bool
-	outputTags      []string
-	hashedImageName string
-	localTag        string
-	dockerUsername  string
-	dockerPassword  string
-	dockerRegistry  string
-	customHost      bool
+	customDockerfile bool
+	customHost       bool
+	dockerPassword   string
+	dockerRegistry   string
+	dockerUsername   string
+	hashExists       bool
+	hashedImageName  string
+	imageHash        string
+	localTag         string
+	outputTags       []string
+	version          string
 }